requirements.txt
Flask==3.0.3
flask_bcrypt==1.0.1
flask_sqlalchemy==3.1.1
PyJWT==2.9.0
